fixed OrderProcessor class and updated the test class, having a mock of ProductRepository

+++ src/main/java/org/example/Product.java
@@ -3,10 +3,10 @@ package org.example;
 public class Product {
     private int id;
     private String name;
-    private int price;
+    private Double price;
 
 
-    public Product(int id, String name, int price) {
+    public Product(int id, String name, Double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -28,11 +28,11 @@ public class Product {
         this.name = name;
     }
 
-    public int getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
 }
